backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "whiteduke",email = "zef1rflis@yandex.ru"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.133.1,<0.134.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.47,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "httpx (>=0.28.0,<1.0.0)",
    "fpdf2 (>=2.8.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)",
    "isort (>=8.0.0,<9.0.0)"
]
